Added code to build a WAV file

+++ src/WaveMachine/Sampling.hs
@@ -2,15 +2,15 @@ module WaveMachine.Sampling where
 
 import Data.Int
 
-sampleTimes :: Double -> Double -> Double -> [Double]
+sampleTimes :: Int -> Double -> Double -> [Double]
 sampleTimes frequency start end
     | start > end = []
-    | otherwise = start:(sampleTimes frequency (start + (1.0 / frequency)) end)
+    | otherwise = start:(sampleTimes frequency (start + (1.0 / (fromIntegral frequency))) end)
 
-sample :: (Double -> Double) -> Double -> Double -> [Double]
+sample :: (Double -> Double) -> Int -> Double -> [Double]
 sample audioFn frequency duration = [audioFn t | t <- sampleTimes frequency 0 duration]
 
-sampleInt16 :: (Double -> Double) -> Double -> Double -> [Int16]
+sampleInt16 :: (Double -> Double) -> Int -> Double -> [Int16]
 sampleInt16 audioFn frequency duration = 
     [floor (v * 32767.5) | v <- samples]
     where samples = sample audioFn frequency duration
